Example: Bump to go1.23 and sriov-network-operator v1.6.0, drop metallb
The example wouldn't compile any more with newer versions of golang.
Bump everything to sriov-network-operator v1.6.0 and golang1.23.

Adjust replace of k8sreporter to use latest local code always.

There was a conflict between flag kubeconfig between metallb and
flag.Parse, meaning that somewhere in metallb flag kubeconfig is now
parsed likely in an init() somewhere - checking diff from v0.10.2
to newer this likely happens in test code. Given that this is only
an example, drop metallb in order to resolve the conflict and to
have a working example.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>